Shaggygoblin patch 1 (#1)
* Some things mentioned in the Forum.

compare to the original 'part.cfg's

* Update part.cfg

This should address the missing LF/OX from the BBQ probe.  As it was, in the editor and r-click popups, there was no fuel for the fire extinguisher showing, and it wouldn't run when attached to the bottom directly and launched.  the resources were 'outside' the "part" module due to bracketing error.

* just saving as .png didn't seem to work, reverting...

* Delete BasicPlywoodFinSmall.png

* Update br_Xe_05b.cfg

forgot to give it a unique name so we don't have to look for the resource on the part to see which one it is

* Update BBQ_3.7.cfg

errant brackets

* Update br_radial_LFOX_tank.cfg

* Update part.cfg